+++ assets/js/DraftList.vue
@@ -3,8 +3,9 @@
         <ul>
             <li v-for="(draft, index) in drafts">
                 <a :href="'/rapport/draft/'+index">
-                    Brouillon n°{{ index }}, mission du {{ draft.metadata.debut }} édité le {{ draft.metadata.edit }}
-                </a>
+                    Brouillon n°{{ index }}
+                </a>, mission du {{ draft.metadata.debut }} édité le {{ draft.metadata.edit }}
+                <a class="button small warning" @click="deleteDraft(index)"><i class="fas fa-trash-alt"></i>Supprimer</a>
             </li>
         </ul>
         <div v-if="1 > drafts.length">Pas de brouillons enregistrés</div>
@@ -20,10 +21,16 @@
         },
         mounted: function() {
             this.drafts = JSON.parse(localStorage.getItem("draft")) || [];
+        },
+        methods: {
+            deleteDraft: function(index) {
+                if(confirm("Cette action efface ce brouillon sur votre poste, elle est définitive. Confirmez-vous la suppression du brouillon ? ")) {
+                    this.drafts.splice(index, 1);
+                    localStorage.setItem("draft", JSON.stringify(this.drafts));
+                }
+            }
         }
     }
 </script>
 
-<style scoped>
-
-</style>
+<style scoped></style>